refactor(provider): extract model catalog domain from tau-coding-agent
Refs #933
