+++ crates/tau-provider/Cargo.toml
@@ -8,9 +8,11 @@ license.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
 base64.workspace = true
+reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
+shell-words.workspace = true
 tau-ai = { path = "../tau-ai" }
 tau-cli = { path = "../tau-cli" }
 tau-core = { path = "../tau-core" }
@@ -18,4 +20,5 @@ tokio.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
+httpmock = "0.8"
 tempfile = "3"
